+++ api/serializers/__init__.py
@@ -3,3 +3,4 @@ from .get_projects_for_obj import get_projects_for_obj
 from .projects_field import ProjectsField
 from .new_threshold_field import NewThresholdField
 from .app_bookmark_field import AppBookmarkField
+from .tag_related_field import TagRelatedField
